Segregate MaybeLight code
